backend/requirements.txt
flask==3.0.0
werkzeug==3.0.1
python-dotenv==1.0.0

# Google Gemini SDK
google-genai==1.55.0

# Database
pymongo==4.6.1

# PDF generation
reportlab==4.0.7

# HTTP
requests==2.31.0
